tests/wpt/tests/payment-handler/payment-request-event-constructor.https.html

<!doctype html>
<meta charset="utf-8">
<title>Test for PaymentRequestEvent Constructor</title>
<link rel="help" href="https://w3c.github.io/payment-handler/#the-paymentrequestevent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>

test(() => {
  assert_false('PaymentRequestEvent' in window);
}, 'PaymentRequestEvent constructor must not be exposed in window');

</script>
